Add next steps of integration loop (incomplete)

+++ engine/src/Symbol/SubexpressionVacancy.cuh
@@ -5,8 +5,8 @@
 
 namespace Sym {
     DECLARE_SYMBOL(SubexpressionVacancy, true)
-    size_t candidate_integral_count;
-    size_t candidate_expression_count;
+    unsigned int candidate_integral_count;
+    unsigned int candidate_expression_count;
     int is_solved; // Nie `bool`, bo nie ma implementacji atomicCAS dla `bool` w CUDA
     size_t solver_idx;
     std::string to_string() const;
